+++ RELEASE_NOTES.md
@@ -1,10 +1,21 @@
 # Release Notes
 
+##Version TBD
+_Date TBD_
+
 ### Minor Changes
 * Change ROLE_ACB_ADMIN to ROLE_ACB
 * Change ROLE_ATL_ADMIN to ROLE_ATL
 * Remove ROLE_ACB_STAFF
 * Remove ROLE_ATL_STAFF
+
+### Table modification
+* Add test_data and test_procedure_temp tables along with criteria mapping tables for each.
+* Change test data version and test procedure version character limits to 50
+* Point certification result tables to new test data and test procedure foreign keys
+
+### Data changes
+* Add v12 2015 upload template
 
 ---